+++ src/clj/coffi/ffi.clj
@@ -269,7 +269,10 @@
   (-> symbol
       ensure-address
       (make-downcall args ret)
-      (make-serde-wrapper args ret)))
+      (cond->
+        (every? #(= % (mem/primitive-type %))
+                (cons ret args))
+        (make-serde-wrapper args ret))))
 
 (defn vacfn-factory
   "Constructs a varargs factory to call the native function referenced by `symbol`.
@@ -548,7 +551,6 @@
         args-types (gensym "args-types")
         ret-type (gensym "ret-type")
         address (gensym "symbol")
-        invoke (gensym "invoke")
         native-sym (gensym "native")
         [arity fn-tail] (-> args :wrapper :fn-tail)
         fn-tail (case arity
@@ -562,14 +564,7 @@
     `(let [~args-types ~(:native-arglist args)
            ~ret-type ~(:return-type args)
            ~address (find-symbol ~(name (:symbol args)))
-           ~invoke (make-downcall ~address ~args-types ~ret-type)
-           ~(or (-> args :wrapper :native-fn) native-sym)
-           ~(if (and (every? #(= % (mem/primitive-type %))
-                             (:native-arglist args))
-                     (= (:return-type args)
-                        (mem/primitive-type (:return-type args))))
-              invoke
-              `(make-serde-wrapper ~invoke ~args-types ~ret-type))
+           ~native-sym (cfn ~address ~args-types ~ret-type)
            fun# ~(if (:wrapper args)
                    `(fn ~(:name args)
                       ~@fn-tail)